feat(delete_remote_older): retention on remote host
This new feature works with INBOX sync.

It allows to set a message retention (expressed in days) for the remote
host. If messages are not expunged immediately after the transfer,
duplicate may be generated: if old messages are retained syncronization
considers only messages with the IMAP \Recent flag. In this case the
remote account must be exclusively accessed by a single Imapsync task.